Add capitalize and trimmer file
