+++ decorator.rb
@@ -1,5 +1,4 @@
 require './nameable'
-require './person'
 
 class BaseDecorator < Nameable
   def initialize(nameable)
@@ -9,17 +8,5 @@ class BaseDecorator < Nameable
 
   def correct_name
     @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name.size > 10 ? @nameable.correct_name[0..9] : @nameable.correct_name
   end
 end
